+++ src/main/java/com/finco/finco/infrastructure/account/dto/AccountRegistrationData.java
@@ -1,5 +1,7 @@
 package com.finco.finco.infrastructure.account.dto;
 
+import java.math.BigDecimal;
+
 import com.finco.finco.entity.account.model.AccountType;
 import com.finco.finco.entity.account.model.CurrencyEnum;
 import com.finco.finco.usecase.account.dto.IAccountRegistrationData;
@@ -8,7 +10,7 @@ public record AccountRegistrationData(
         Long userId,
         String name,
         AccountType type,
-        Long balance,
+        BigDecimal balance,
         CurrencyEnum currency,
         String description) implements IAccountRegistrationData {
