+++ src/main/java/com/finco/finco/usecase/account/dto/IAccountTransactionData.java
@@ -1,10 +1,12 @@
 package com.finco.finco.usecase.account.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.Positive;
 
 public interface IAccountTransactionData {
 
     @Positive(message = "Amount must be greater than zero")
-    Long amount();
+    BigDecimal amount();
 
 }
